Let's just forget about uints

+++ AetherRadio.Numerics.Tests/FunctionsTests.cs
@@ -6,12 +6,12 @@ public class FunctionsTests
     [Test]
     public void IsPowerOfTwoShouldReturnTrueForPowersOfTwo([Range(0, 16)] int x)
     {
-        Assert.That(Functions.IsPowerOfTwo((uint)Math.Pow(2, x)));
+        Assert.That(Functions.IsPowerOfTwo((int)Math.Pow(2, x)));
     }
 
     [Test]
     public void IsPowerOfTwoShouldReturnFalseForNonPowersOfTwo([Values(0, 3, 5, 6, 7, 9, 10, 11, 12, 13, 15, 17, 31, 100, 1000)] int x)
     {
-        Assert.That(Functions.IsPowerOfTwo((uint)x), Is.False);
+        Assert.That(Functions.IsPowerOfTwo(x), Is.False);
     }
 }
